Switch providers to a GRPC backend.
To make it easier to use RUN-DSP as a dataspace connector, we're moving
away from needing integrators to write go code. Instead they can
write the provider in any language and it will work with RUN-DSP as long
as they use the GRPC bindings.

What this PR contains:

- Removing the code for the filesystem test provider.
- Configuration method for how to connect to a provider backend.
- Moved the provider calls to GRPC.

What is still to do:

- Forward the auth header to the provider.
- Get some information for the dataservice heading.