Add contract hooks.
This PR adds a GRPC client for the contract service while also
adding control functions to send contract negotiation messages.
This makes it possible to have fine grained control over contract
negotiation approval. This is lacking the termination implementation
but that will be implemented soon in a separate PR.

Changes:

- Split up receive and send steps in state machine.
- Only do receive steps in http handlers.
- Add new control GRPC methods in charge of sending steps.
- Add new mocks for contract service.
- Upgrade GRPC bindings.
- Update mockery.
- Add token verification and storage for the token.