Allow auto accept on contract negotiations.
This allows RUN-DSP to be configured without a contract service,
or for a contract service to set a negotiation to "auto accept".

When RUN-DSP handles a negotiation that's set to auto accept, it will
automatically accept and continue to the next step of the negotiation.
Removing the need for any input from the contract service from that
point forward.